# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SenAgri B2B",
        "description": "Documentation officielle de l'API SenAgri B2B",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/audit-logs": {
            "get": {
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Lister les journaux d'audit",
                "description": "Retourne la liste paginée de tous les journaux d'audit.",
                "operationId": "getAuditLogs",
                "responses": {
                    "200": {
                        "description": "Liste des journaux récupérée avec succès"
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    },
                    "403": {
                        "description": "Accès interdit"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/audit-logs/{id}": {
            "get": {
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Afficher un journal d'audit",
                "description": "Retourne les informations d'un journal d'audit précis.",
                "operationId": "showAuditLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du journal d'audit",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Journal récupéré avec succès"
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    },
                    "404": {
                        "description": "Journal introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/commandes": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Lister les commandes",
                "operationId": "63413d0f1ac9da2056a385b55f791e37",
                "responses": {
                    "200": {
                        "description": "Liste des commandes"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Créer une commande",
                "operationId": "5d962071e01ba227e33a4b0122eaeced",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "offre_id",
                                    "quantite"
                                ],
                                "properties": {
                                    "offre_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantite": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commande créée avec succès"
                    },
                    "403": {
                        "description": "Impossible de commander sa propre offre"
                    },
                    "422": {
                        "description": "Offre indisponible ou stock insuffisant"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/commandes/{id}": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Afficher une commande",
                "operationId": "387ffac66f5018f15c418bc864327bfc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de la commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commande trouvée"
                    },
                    "404": {
                        "description": "Commande introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Modifier le statut d'une commande",
                "operationId": "4bfe75d25e7159c22ea4d21f46b32d7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de la commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "statut"
                                ],
                                "properties": {
                                    "statut": {
                                        "type": "string",
                                        "example": "confirmee"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commande mise à jour"
                    },
                    "404": {
                        "description": "Commande introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Supprimer une commande",
                "operationId": "bb024d05486cc9cc40b761a34137a04d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de la commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commande supprimée"
                    },
                    "404": {
                        "description": "Commande introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}